Update type for map
Update version

+++ src/main/java/de/floribe2000/warships_java/direct/warships/Statistics.java
@@ -27,7 +27,7 @@ public class Statistics implements IApiResponse {
 
 	private Meta meta = null;
 
-	private Map<Integer, List<ShipEntry>> data = null;
+	private Map<Long, List<ShipEntry>> data = null;
 
 	@Getter
 	public static class ShipEntry {
